+++ src/redturtle/exporter/base/browser/jsonify.py
@@ -4,13 +4,12 @@ from .wrapper import Wrapper
 from DateTime import DateTime
 from plone import api
 from plone.app.discussion.interfaces import IConversation
-from ploneorg.jsonify.jsonify import GetItem as BaseGetItemView
-from ploneorg.jsonify.jsonify import GetCatalogResults as BaseGetCatalogResults
 from Products.CMFCore.interfaces import ISiteRoot
+from Products.Five.browser import BrowserView
 
 # navigation tree
 from Products.CMFCore.interfaces import IFolderish
-from Acquisition import aq_parent
+from Products.CMFCore.interfaces import ISiteRoot
 
 import base64
 import json
@@ -111,9 +110,42 @@ def get_taxonomy_object(self, context_dict):
         del context_dict['siteAreas']
 
 
+class BaseGetItemView(BrowserView):
+
+    def __call__(self):
+        try:
+            context_dict = Wrapper(self)
+        except Exception, e:
+            etype = sys.exc_info()[0]
+            tb = pprint.pformat(traceback.format_tb(sys.exc_info()[2]))
+            return 'ERROR: exception wrapping object: %s: %s\n%s' % (
+                etype, str(e), tb
+            )
+
+        passed = False
+        while not passed:
+            try:
+                JSON = json.dumps(context_dict)
+                passed = True
+            except Exception, error:
+                if "serializable" in str(error):
+                    key, context_dict = _clean_dict(context_dict, error)
+                    pprint.pprint(
+                        'Not serializable member %s of %s ignored' % (
+                            key, repr(self)
+                        )
+                    )
+                    passed = False
+                else:
+                    return ('ERROR: Unknown error serializing object: %s' % error)
+        self.REQUEST.response.setHeader("Content-type", "application/json")
+        return JSON
+
+
 class BaseGetItem(BaseGetItemView):
 
     def __call__(self):
+
         context_dict = Wrapper(self.context)
 
         # funzioni comuni a tutti i get_item
@@ -293,6 +325,9 @@ class GetItemCollection(BaseGetItem):
             tb = pprint.pformat(traceback.format_tb(sys.exc_info()[2]))
             return 'ERROR: exception wrapping object: %s\n%s' % (str(e), tb)
 
+        context_dict['item_count'] = context_dict.get('limit', 30)
+        del context_dict['limit']
+
         return get_json_object(self, context_dict)
 
 
@@ -334,7 +369,7 @@ class GetItemImage(BaseGetItem):
         return get_json_object(self, context_dict)
 
 
-class GetCatalogResults(BaseGetCatalogResults):
+class GetCatalogResults(object):
 
     items = []
     item_paths = []
@@ -342,8 +377,10 @@ class GetCatalogResults(BaseGetCatalogResults):
     def flatten(self, children):
         """ Recursively flatten the tree """
         for obj in children:
-            self.items.append(obj["path"])
-            children = obj.get("children", None)
+            if obj['path']:
+                self.items.append(obj['path']) 
+
+            children = obj.get('children', None)
             if children:
                 self.flatten(children)
 
@@ -384,15 +421,16 @@ class GetCatalogResults(BaseGetCatalogResults):
                 if not root:
                     return json.dumps(self.item_paths)
 
-                if not IFolderish.providedBy(root):
-                    self.item_paths.append(root.absolute_url_path())
-                    return self.item_paths
+                # if not IFolderish.providedBy(root):
+                #     self.items.append(root.absolute_url_path())
+                #     return json.dumps(self.items)
 
                 path = root.absolute_url_path() if not getattr(root, "getObject", None) else root.getPath() # noqa
-                tree = {'path': path, 'children': []}
+                tree = {'path': path, 'children': []} if not ISiteRoot.providedBy(root) else {'children': []} # noqa
                 tree['children'].extend(self.explain_tree(root))
 
-                self.items.append(tree['path'])
+                if tree.get('path', None):
+                    self.items.append(tree['path'])
 
                 self.flatten(tree['children'])
                 item_paths = self.items